feat(cli): serve worktrees from a headless host
`oximux serve` wired no worktree service, so an authorized caller got
Unsupported from `worktree create/ls/rm`, `run --worktree`, and
`team run --worktree-each` — most of the reason to run a headless host in the
first place. The only implementation lived in the desktop, reachable only
through a gpui-importing module.

Extract the lifecycle into `oximux-worktree-ops`, following relay-terminals:
the path scheme, create-with-rollback, the pre-remove cleanup, and the
WorktreeService impl, all gpui-free. Both hosts now run the same one, so a
worktree either creates is the row the other lists.

Two things the move forced:

- `worktree_path` takes the data dir as an argument instead of resolving the
  desktop's app-data root. Resolving it internally would put a server's
  worktrees under the desktop's directory.
- serve registers its `--project` roots as `projects` rows at boot, since
  creation resolves a client-named root to a project id and serve held only
  paths. `insert_or_touch` is idempotent, so this is safe every boot and
  against the desktop having already added the same path.

`load_for_project` moves to oximux-settings beside the type it parses, with
its tests; the desktop keeps its own name for it as a re-export.

Covered by a serve_e2e case that drives the real binary and checks git rather
than the reply — the existing worktree tests build a dispatcher with a stub
service and could not have caught this.
