+++ crates/settings/Cargo.toml
@@ -21,6 +21,7 @@ serde = { workspace = true }
 toml = { workspace = true }
 anyhow = { workspace = true }
 dunce = { workspace = true }
+tracing = { workspace = true }
 
 [dev-dependencies]
 # Only to give the project-coverage tests a real directory to canonicalize —
